PRISM_backend/requirements.txt
google-cloud-storage>=2.0.0
fastapi
anyio==4.9.0
SQLAlchemy==2.0.40
pydantic-settings==2.8.1
passlib==1.7.4
tensorflow==2.19.0
torch==2.6.0
transformers==4.50.3
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.6.1
reportlab==4.3.1
matplotlib==3.10.1
shap==0.47.1
lime==0.2.0.1
psutil==7.0.0
imbalanced-learn==0.13.0
python-multipart==0.0.20
seaborn==0.13.2
uvicorn==0.34.0
bcrypt
numpy==1.26.4
alembic==1.15.2
supabase
pydantic[email]==2.8.1
gunicorn
python-dotenv==1.0.0
python-jose>=3.3.0
pydantic_settings